Feature/osa4

See merge request vhiltsu/fs-open!52

+++ osa4/blogilista/controllers/blogs.js
@@ -1,13 +1,24 @@
-const blogsRouter = require('express').Router()
-const { Blog } = require('../models/blog')
+const blogsRouter = require("express").Router();
+const Blog = require("../models/blog");
+const { userExtractor } = require("../utils/middleware");
 
-blogsRouter.get("", async (request, response) => {
-  const blog = await Blog.find({});
+blogsRouter.get("/", async (request, response) => {
+  const blog = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+    id: 1,
+  });
   response.json(blog);
 });
 
-blogsRouter.post("", async (request, response) => {
-  const blog = new Blog(request.body);
+blogsRouter.post("/", userExtractor, async (request, response) => {
+  const user = request.user;
+
+  if (!user) {
+    return response.status(400).json({ error: "UserId missing or not valid" });
+  }
+
+  const blog = new Blog({ ...request.body, user: user._id });
 
   // Validate the blog data
   try {
@@ -18,24 +29,35 @@ blogsRouter.post("", async (request, response) => {
 
   // Save the new blog to the database
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
   // Attempt to delete the blog with the given ID
-  try {
-    await Blog.findByIdAndDelete(request.params.id);
-    response.status(204).end();
-  } catch (error) {
-    response.status(404).end();
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
+    return response.status(404).end();
   }
+
+  if (blog.user.toString() !== request.user.id) {
+    return response
+      .status(401)
+      .json({ error: "only the creator can delete a blog" });
+  }
+
+  await Blog.findByIdAndDelete(request.params.id);
+
+  response.status(204).end();
 });
 
 blogsRouter.put("/:id", async (request, response) => {
   // Check if the blog with the given ID exists
   try {
     await Blog.findById(request.params.id);
-  } catch (error) {
+  } catch {
     return response.status(404).end();
   }
 
@@ -46,12 +68,15 @@ blogsRouter.put("/:id", async (request, response) => {
     return response.status(400).json({ error: error.message });
   }
 
+  delete request.body.id;
+
   const updatedBlog = await Blog.findByIdAndUpdate(
     request.params.id,
     request.body,
-    { new: true, runValidators: true, context: "query" }
+    { returnDocument: "after", runValidators: false }
   );
+
   response.json(updatedBlog);
 });
 
-module.exports = blogsRouter
+module.exports = blogsRouter;
